+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/memory/SimpleByteArraySupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -48,6 +48,8 @@ package com.oracle.truffle.api.memory;
  */
 @SuppressWarnings("PointlessArithmeticExpression")
 final class SimpleByteArraySupport extends ByteArraySupport {
+    private final SimpleByteArraySupportLock lock = new SimpleByteArraySupportLock();
+
     @Override
     public byte getByte(byte[] buffer, int byteOffset) throws IndexOutOfBoundsException {
         return buffer[byteOffset];
@@ -200,5 +202,285 @@ final class SimpleByteArraySupport extends ByteArraySupport {
     public void putDouble(byte[] buffer, long byteOffset, double value) throws IndexOutOfBoundsException {
         assert byteOffset < Integer.MAX_VALUE;
         putDouble(buffer, (int) byteOffset, value);
+    }
+
+    @Override
+    public byte getByteVolatile(byte[] buffer, long byteOffset) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            return getByte(buffer, byteOffset);
+        }
+    }
+
+    @Override
+    public void putByteVolatile(byte[] buffer, long byteOffset, byte value) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            putByte(buffer, byteOffset, value);
+        }
+    }
+
+    @Override
+    public short getShortVolatile(byte[] buffer, long byteOffset) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            return getShort(buffer, byteOffset);
+        }
+    }
+
+    @Override
+    public void putShortVolatile(byte[] buffer, long byteOffset, short value) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            putShort(buffer, byteOffset, value);
+        }
+    }
+
+    @Override
+    public int getIntVolatile(byte[] buffer, long byteOffset) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            return getInt(buffer, byteOffset);
+        }
+    }
+
+    @Override
+    public void putIntVolatile(byte[] buffer, long byteOffset, int value) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            putInt(buffer, byteOffset, value);
+        }
+    }
+
+    @Override
+    public long getLongVolatile(byte[] buffer, long byteOffset) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            return getLong(buffer, byteOffset);
+        }
+    }
+
+    @Override
+    public void putLongVolatile(byte[] buffer, long byteOffset, long value) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            putLong(buffer, byteOffset, value);
+        }
+    }
+
+    @Override
+    public byte getAndAddByte(byte[] buffer, long byteOffset, byte delta) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            putByte(buffer, byteOffset, (byte) (v + delta));
+            return v;
+        }
+    }
+
+    @Override
+    public short getAndAddShort(byte[] buffer, long byteOffset, short delta) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            putShort(buffer, byteOffset, (short) (v + delta));
+            return v;
+        }
+    }
+
+    @Override
+    public int getAndAddInt(byte[] buffer, long byteOffset, int delta) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            putInt(buffer, byteOffset, v + delta);
+            return v;
+        }
+    }
+
+    @Override
+    public long getAndAddLong(byte[] buffer, long byteOffset, long delta) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            putLong(buffer, byteOffset, v + delta);
+            return v;
+        }
+    }
+
+    @Override
+    public byte getAndBitwiseAndByte(byte[] buffer, long byteOffset, byte mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            putByte(buffer, byteOffset, (byte) (v & mask));
+            return v;
+        }
+    }
+
+    @Override
+    public short getAndBitwiseAndShort(byte[] buffer, long byteOffset, short mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            putShort(buffer, byteOffset, (short) (v & mask));
+            return v;
+        }
+    }
+
+    @Override
+    public int getAndBitwiseAndInt(byte[] buffer, long byteOffset, int mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            putInt(buffer, byteOffset, v & mask);
+            return v;
+        }
+    }
+
+    @Override
+    public long getAndBitwiseAndLong(byte[] buffer, long byteOffset, long mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            putLong(buffer, byteOffset, v & mask);
+            return v;
+        }
+    }
+
+    @Override
+    public byte getAndBitwiseOrByte(byte[] buffer, long byteOffset, byte mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            putByte(buffer, byteOffset, (byte) (v | mask));
+            return v;
+        }
+    }
+
+    @Override
+    public short getAndBitwiseOrShort(byte[] buffer, long byteOffset, short mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            putShort(buffer, byteOffset, (short) (v | mask));
+            return v;
+        }
+    }
+
+    @Override
+    public int getAndBitwiseOrInt(byte[] buffer, long byteOffset, int mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            putInt(buffer, byteOffset, v | mask);
+            return v;
+        }
+    }
+
+    @Override
+    public long getAndBitwiseOrLong(byte[] buffer, long byteOffset, long mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            putLong(buffer, byteOffset, v | mask);
+            return v;
+        }
+    }
+
+    @Override
+    public byte getAndBitwiseXorByte(byte[] buffer, long byteOffset, byte mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            putByte(buffer, byteOffset, (byte) (v ^ mask));
+            return v;
+        }
+    }
+
+    @Override
+    public short getAndBitwiseXorShort(byte[] buffer, long byteOffset, short mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            putShort(buffer, byteOffset, (short) (v ^ mask));
+            return v;
+        }
+    }
+
+    @Override
+    public int getAndBitwiseXorInt(byte[] buffer, long byteOffset, int mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            putInt(buffer, byteOffset, v ^ mask);
+            return v;
+        }
+    }
+
+    @Override
+    public long getAndBitwiseXorLong(byte[] buffer, long byteOffset, long mask) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            putLong(buffer, byteOffset, v ^ mask);
+            return v;
+        }
+    }
+
+    @Override
+    public byte getAndSetByte(byte[] buffer, long byteOffset, byte newValue) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            putByte(buffer, byteOffset, newValue);
+            return v;
+        }
+    }
+
+    @Override
+    public short getAndSetShort(byte[] buffer, long byteOffset, short newValue) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            putShort(buffer, byteOffset, newValue);
+            return v;
+        }
+    }
+
+    @Override
+    public int getAndSetInt(byte[] buffer, long byteOffset, int newValue) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            putInt(buffer, byteOffset, newValue);
+            return v;
+        }
+    }
+
+    @Override
+    public long getAndSetLong(byte[] buffer, long byteOffset, long newValue) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            putLong(buffer, byteOffset, newValue);
+            return v;
+        }
+    }
+
+    @Override
+    public byte compareAndExchangeByte(byte[] buffer, long byteOffset, byte expected, byte x) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            byte v = getByte(buffer, byteOffset);
+            if (v == expected) {
+                putByte(buffer, byteOffset, x);
+            }
+            return v;
+        }
+    }
+
+    @Override
+    public short compareAndExchangeShort(byte[] buffer, long byteOffset, short expected, short x) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            short v = getShort(buffer, byteOffset);
+            if (v == expected) {
+                putShort(buffer, byteOffset, x);
+            }
+            return v;
+        }
+    }
+
+    @Override
+    public int compareAndExchangeInt(byte[] buffer, long byteOffset, int expected, int x) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            int v = getInt(buffer, byteOffset);
+            if (v == expected) {
+                putInt(buffer, byteOffset, x);
+            }
+            return v;
+        }
+    }
+
+    @Override
+    public long compareAndExchangeLong(byte[] buffer, long byteOffset, long expected, long x) throws IndexOutOfBoundsException {
+        synchronized (lock) {
+            long v = getLong(buffer, byteOffset);
+            if (v == expected) {
+                putLong(buffer, byteOffset, x);
+            }
+            return v;
+        }
     }
 }