+++ truffle/CHANGELOG.md
@@ -31,6 +31,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 	* `engine.CompilationExceptionsAreFatal`: Use `engine.CompilationFailureAction=ExitVM` instead.
 * GR-44420 Added `TruffleLanguage.finalizeThread(Object, Thread)` to allow languages run finalization hooks for initialized threads before the context is disposed.
 * GR-45923 Added `EventBinding.tryAttach()` to try to attach a binding, if not disposed or attached already.
+* GR-20628 Added atomic byte-array operations to `ByteArraySupport` and subclasses.
 * GR-39571 Added `TranscodingErrorHandler` to `TruffleString.SwitchEncodingNode`. 
 * GR-46345 Added a support for the lazy unpacking of language and instrument resources necessary for execution. This support replaces the concept of language homes for Maven language and tool deployment. For a language or instrument that requires additional files to execute, it needs to follow these steps:
   * Bundle the necessary files into a jar distribution.